# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Dijigoo Kurye Mobil API
  version: 1.0.0-draft.2
  description: |-
    Dijigoo Kurye saha uygulamasinin mobil API sozlesmesi.

    Iki urun, bir kimlik: web (kurye.dijigoo.com) onboarding tasir;
    mobil saha operasyonunu tasir. Ayni Next.js ailesinde yasir.

    ## Namespace

    - Mobil token API: `https://kurye.dijigoo.com/api/mobile` + bu belgedeki `/v1/*`
      (tam yol `/api/mobile/v1/...`).
    - Web cookie session (`dijigoo_courier_session`) dokunulmaz; mobil kullanmaz.
    - Web public API (`/api/public/v1/courier-*`) kimlik kaynagi olarak devralinir.
    - Iki OTP havuzu: kurye aktivasyon (`courierId`) ≠ alici teslim (`taskId+phone`).

    ## Sozlesme kurallari

    - Tum tarih alanlari RFC 3339, UTC ve acik offsetli.
    - Tum mutasyon istekleri `Idempotency-Key` basligi tasir. Ayni anahtarla
      gelen tekrar istegi ilk yanitin aynisini doner.
    - Gorev mutasyonlari `rowVersion` tasir; uyusmazlik 409 VERSION_MISMATCH.
    - Hata govdesi her zaman `application/problem+json` ve `ErrorResponse` semasidir.
    - Sayfalama imlec tabanlidir; `nextCursor` null ise son sayfadir.
    - Surumleme yol tabanlidir (`/v1`). Kirici degisiklik yeni yol acar.
    - Workflow adim tipleri kapali enum. v1 renderer: PHOTO_EVIDENCE ve FORM.select.
    - Vardiya fotografi yuz eslestirme yapmaz (KVKK m.6); `shiftFaceMatch=false`.
  contact:
    name: Dijigoo teknik ekip
servers:
  - url: https://kurye.dijigoo.com/api/mobile
...
    description: Yerel panel
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Aktivasyon sonrasi verilen kisa omurlu access token. 15 dakika gecerli, refresh ile
        doner. Token cihaz kimligine (installationId) baglidir.
    serviceAuth:
      type: apiKey
      in: header
      name: x-service-token
      description: Operasyon panelinin backend’i icin paylasilan servis anahtari. Kurye bearer tokeni
        degildir, bu API’de operator kimligi yoktur.
    solvelineInboundAuth:
      type: http
      scheme: bearer
      description: Solveline sesli asistanin bizim IVR sorgu/ticket API’sini cagirmasi icin ayri servis
        tokeni (SOLVELINE_INBOUND_TOKEN). Kurye JWT ve panel x-service-token degildir.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: &a17
                - VALIDATION_FAILED
                - MALFORMED_REQUEST
                - UNSUPPORTED_CLIENT_VERSION
                - UNAUTHENTICATED
                - TOKEN_EXPIRED
                - TOKEN_REVOKED
                - OTP_INVALID
                - OTP_EXPIRED
                - FORBIDDEN
                - DEVICE_NOT_BOUND
                - DEVICE_INTEGRITY_FAILED
                - SHIFT_NOT_ACTIVE
                - GEOFENCE_VIOLATION
                - WORKFLOW_STEP_OUT_OF_ORDER
                - NOT_FOUND
                - CONFLICT
                - VERSION_MISMATCH
                - IDEMPOTENCY_KEY_REUSED
                - TASK_ALREADY_FINALIZED
                - WORKFLOW_VERSION_SUPERSEDED
                - MEDIA_NOT_READY
                - PAYLOAD_TOO_LARGE
                - UNSUPPORTED_MEDIA_TYPE
                - BUSINESS_RULE_VIOLATION
                - EVIDENCE_INCOMPLETE
                - RATE_LIMITED
                - INTERNAL_ERROR
                - UPSTREAM_UNAVAILABLE
                - UPSTREAM_TIMEOUT
            message:
              type: string
            userVisible:
              type: boolean
              default: false
            traceId:
              type: string
              example: 0af7651916cd43dd8448eb211c80319c
            details:
              type:
                - array
                - "null"
              items:
                $ref: "#/components/schemas/ErrorDetail"
            retryAfter:
              type:
                - integer
                - "null"
              exclusiveMinimum: 0
          required:
            - code
            - message
            - traceId
      required:
        - error
...
    CustodyListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CustodyItem"
        nextCursor:
          type:
            - string
            - "null"
          description: null ise son sayfa
        syncedAt:
          type: string
          format: date-time
          example: 2026-08-25T09:15:00.000Z
      required:
        - items
        - nextCursor
        - syncedAt
...
paths:
...
  /v1/custody:
    get:
      tags:
        - Custody
      summary: Kurye uzerindeki zimmet
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type:
              - string
              - "null"
            description: Onceki yanittaki nextCursor
          required: false
          description: Onceki yanittaki nextCursor
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          required: false
          name: limit
          in: query
        - schema:
            type:
              - array
              - "null"
            items:
              type: string
              enum: *a21
          required: false
          name: type
          in: query
        - schema:
            type:
              - string
              - "null"
            maxLength: 80
          required: false
          name: barcode
          in: query
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
      responses:
        "200":
          description: Zimmet listesi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustodyListResponse"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/custody/handover:
    post:
      tags:
        - Custody
      summary: Zimmet devret veya devral
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
        - schema:
            type: string
            format: uuid
            description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
              istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          required: true
          description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
            istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          name: idempotency-key
          in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustodyHandoverRequest"
      responses:
        "200":
          description: Devir tamam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustodyHandoverResponse"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Surum catismasi veya kayit zaten sonuclandirilmis
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Is kurali ihlali veya eksik kanit
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/support/tickets:
    get:
      tags:
        - Support
      summary: Destek kayitlari
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type:
              - string
              - "null"
            description: Onceki yanittaki nextCursor
          required: false
          description: Onceki yanittaki nextCursor
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          required: false
          name: limit
          in: query
        - schema:
            type:
              - string
              - "null"
            maxLength: 80
          required: false
          name: barcode
          in: query
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
      responses:
        "200":
          description: Kayit listesi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketListResponse"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Support
      summary: Destek kaydi ac
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
        - schema:
            type: string
            format: uuid
            description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
              istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          required: true
          description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
            istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          name: idempotency-key
          in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketCreateRequest"
      responses:
        "201":
          description: Kayit acildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Surum catismasi veya kayit zaten sonuclandirilmis
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Is kurali ihlali veya eksik kanit
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...